+++ src/components/Collections/CollectionCardWithActiveIndicator.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { EnteFile } from 'types/file';
-import { CollectionTileWrapper, ActiveIndicator } from '.';
+import { CollectionTileWrapper, ActiveIndicator } from './styledComponents';
 import CollectionCard from './CollectionCard';
 
 const CollectionCardWithActiveIndicator = React.forwardRef(
