+++ src/components/Collections/CollectionNamer.tsx
